Factorize db and cache all recorders
